requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
babel==2.17.0
binaryornot==0.4.4
boto3==1.35.71
botocore==1.35.71
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
coverage==7.6.8
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.11.1
defusedxml==0.8.0rc2
Django==5.1.3
django-allauth==0.60.1
django-browser-reload==1.17.0
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-money==3.5.3
django-storages==1.14.4
django-tailwind==3.8.0
djangorestframework==3.16.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
executing==2.1.0
factory_boy==3.3.1
Faker==33.1.0
icecream==2.1.3
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
minio==7.2.12
more-itertools==10.5.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
premailer==3.10.0
psycopg2-binary==2.9.10
py-moneyed==3.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.1
s3transfer==0.10.4
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
text-unidecode==1.3
tomli==2.2.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.2.3
djangorestframework-simplejwt==5.5.0
dj-rest-auth==7.0.1
